Update README.md
Hi, @ExE-Boss. I wanted to talk to you about whether or not you might be interested in being nominated to be a Node.js Collaborator, but I don't see an obvious way to get in touch with you. So I'm opening this pull request just to leave you a note: Can you email me (rtrott@gmail.com) or suggest some other way I might talk privately with you? Or we can have the conversation here (in this PR) if you'd prefer that. Thanks in advance.